:bug: Fix type of BrewTime in brewing_stand

+++ src/data/JE1.15/nbt_schema.ts
@@ -181,7 +181,7 @@ export default {
 				"description": "Items in the brewing stand"
 			},
 			"BrewTime": {
-				"type": "int",
+				"type": "short",
 				"description": "The number of ticks left in brewing"
 			},
 			"Fuel": {
